requirements.txt
# Core agent framework
google-adk==0.5.3          # Agent Development Kit (agents, runner, sessions)

# Google Generative AI SDK (Gemini models)
google-generativeai==0.5.2 # provides google.genai.types.Content etc.

# MCP client used by flights / hotels tools
mcp==1.2.1                 # pip install mcp

# Data validation / schema
pydantic==2.6.4

# Async helpers used under the hood by ADK & MCP
anyio==4.3.0

# Terminal CLI colours
colorama==0.4.7

# .env management
python-dotenv==1.0.1

# Front-end
streamlit==1.34.0

# Optional but common utilities
requests>=2.32.0           # if your tools make HTTP calls
pytest>=8.2.0              # test harness (tests/test_full_stack.py)

# If you pin Python version in tools/CI, document it here:
# python_version >= "3.10"